requirements.txt
streamlit
pandas
numpy
Pillow
plotly
requests